Update new url after moving stock page out
